+++ port/cortex-m0/src/peripheral/gpio/bsrr.rs
@@ -7,19 +7,19 @@ use super::super::Register;
 
 #[derive(Copy, Clone)]
 pub struct BSRR {
-  base_addr: u32,
+  base_addr: usize,
 }
 
 impl Register for BSRR {
-  fn new(base_addr: u32) -> Self {
+  fn new(base_addr: usize) -> Self {
     BSRR { base_addr: base_addr }
   }
 
-  fn base_addr(&self) -> u32 {
+  fn base_addr(&self) -> usize {
     self.base_addr
   }
 
-  fn mem_offset(&self) -> u32 {
+  fn mem_offset(&self) -> usize {
     0x18
   }
 }
