+++ port/cortex-m0/src/peripheral/gpio/moder.rs
@@ -14,7 +14,7 @@ pub enum Mode {
 }
 
 impl Field for Mode {
-  fn mask(&self) -> u32 {
+  fn mask(&self) -> usize {
     match *self {
       Mode::Input => 0b00,
       Mode::Output => 0b01,
@@ -25,7 +25,7 @@ impl Field for Mode {
 }
 
 impl Mode {
-  fn from_mask(mask: u32) -> Self {
+  fn from_mask(mask: usize) -> Self {
     match mask {
       0b00 => Mode::Input,
       0b01 => Mode::Output,
@@ -38,19 +38,19 @@ impl Mode {
 
 #[derive(Copy, Clone)]
 pub struct MODER {
-  base_addr: u32,
+  base_addr: usize,
 }
 
 impl Register for MODER {
-  fn new(base_addr: u32) -> Self {
+  fn new(base_addr: usize) -> Self {
     MODER { base_addr: base_addr }
   }
 
-  fn base_addr(&self) -> u32 {
+  fn base_addr(&self) -> usize {
     self.base_addr
   }
 
-  fn mem_offset(&self) -> u32 {
+  fn mem_offset(&self) -> usize {
     0x0
   }
 }
